ver. 0.11.3.0 Changed the argument datatype and return type of methods size_in_bytes(), size_in_bits(), length_in_bytes(), length_in_bits(), submax(), set_submax(), is_bit_set(), is_bit_set_(), generate_check_bits(), generate_check_bits_(), and get_lower_portion() of BigUInt from usize into u32, and changed the datatype of constant BITS and the argument datatype and return type of methods size_in_bytes(), size_in_bits(), length_in_bytes(), length_in_bits(), set_submax(), is_bit_set(), is_bit_set_(), generate_check_bits(), and generate_check_bits_() of SmallUInt from usize into u32, and changed the argument datatype of methods digest(), and ruminate() of SHA_2_512_Generic and SHA_2_512_t_Generic from u128 into u64, and removed the methods digest_c(), and ruminate_c() of SHA_2_512_Generic and SHA_2_512_t_Generic, and fixed all the corresponding examples according to the changes.

+++ src/number/macros_for_integer_unions.rs
@@ -10936,7 +10936,7 @@ macro_rules! integer_union_methods {
             let mut low = zero;
             let mut high = zero;
             let mut overflow: bool;
-            let mut bit_check = one << (Self::size_in_bits() - 1 - rhs.leading_zeros() as usize);
+            let mut bit_check = one << (Self::size_in_bits() - 1 - rhs.leading_zeros());
             let adder = self;
             while !bit_check.is_zero()
             {
